# boot/assets/sw/config/swagger.yaml
...
info:
  contact:
    email: lark@pointgoal.io
    name: rk-dev
    url: https://github.com/rookie-ninja/rk-fiber
  description: This is builtin RK common service.
  license:
    name: Apache 2.0 License
    url: https://github.com/rookie-ninja/rk-fiber/blob/master/LICENSE.txt
  title: RK Common Service
  version: "1.0"
paths:
...
  /rk/v1/tv/:
    get:
      operationId: "15"
      produces:
      - text/html
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      - BasicAuth: []
      summary: Get HTML page of /tv/
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
  BasicAuth:
    type: basic
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"